Dump more messages for tick overflow or round start
E.g., try with "jitterz -d 1" and I see it'll run actually ~3sec.  It's because
we've got a retry because hz diff too big.  Dump all these information.

$ time sudo ./jitterz -d 1
Running with expected frequency: 4800.000000 (MHz)
Running with expected frequency: 2112.018929 (MHz)
...

Also dump something on tick overflow so in case (though I doubt it!) if it
happens we also know why we've retried.

Signed-off-by: Peter Xu <peterx@redhat.com>